src/workflow.py
from langgraph.graph import START, StateGraph
from langchain_core.documents import Document
from typing import List, TypedDict


class State(TypedDict):
    question: str
    context: List[Document]
    answer: str


def build_workflow(vector_store, llm, prompt):
    def retrieve(state: State):
        retrieved_docs = vector_store.similarity_search(state["question"])
        return {"context": retrieved_docs}

    def generate(state: State):
        docs_content = "\n\n".join(doc.page_content for doc in state["context"])
        messages = prompt.invoke(
            {"question": state["question"], "context": docs_content}
        )
        response = llm.invoke(messages)
        return {"answer": response.content}

    graph_builder = StateGraph(State).add_sequence([retrieve, generate])
    graph_builder.add_edge(START, "retrieve")
    return graph_builder.compile()
