pyproject.toml
[project]
name = "evergreen-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "langchain>=0.3.13",
    "langchain-community>=0.3.13",
    "langchain-openai>=0.2.14",
    "langchain-text-splitters>=0.3.4",
    "langgraph>=0.2.60",
    "python-dotenv>=1.0.1",
    "langchain-huggingface>=0.1.2",
    "langchain-chroma>=0.1.4",
    "unstructured[md]>=0.16.11",
    "nltk>=3.9.1",
    "discord-py>=2.4.0",
]
